requirements.txt
streamlit>=1.40.0
pandas>=2.2.0
requests>=2.31.0
google-genai>=1.63.0
fake-useragent>=2.2.0
tenacity>=9.0.0
